requirements.txt
# 核心依赖
numpy>=1.20.0
psycopg[binary,pool]>=3.1.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
colorama>=0.4.6
# API 相关依赖
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0 